Documentatie EducomOpdrachtAPI

+++ EducomOpdrachtAPI/Controllers/WeerberichtenController.cs
@@ -14,6 +14,7 @@ namespace EducomOpdrachtAPI.Controllers
     [ApiController]
     public class WeerberichtenController : ControllerBase
     {
+        // Stelt de database context in (Entity Framework)
         private readonly WeerberichtContext _context;
 
         public WeerberichtenController(WeerberichtContext context)
@@ -22,6 +23,10 @@ namespace EducomOpdrachtAPI.Controllers
         }
 
         // GET: api/Weerberichten
+        /// <summary>
+        /// Haalt alle weerberichten op uit de database.
+        /// </summary>
+        /// <returns>Lijst met weerberichten</returns>
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Weerbericht>>> GetWeerberichten()
         {
@@ -29,11 +34,18 @@ namespace EducomOpdrachtAPI.Controllers
         }
 
         // GET: api/Weerberichten/5
+        /// <summary>
+        /// Haalt een specifiek weerbericht op.
+        /// </summary>
+        /// <param name="id">ID van weerbericht</param>
+        /// <returns>Weerbericht</returns>
         [HttpGet("{id}")]
         public async Task<ActionResult<Weerbericht>> GetWeerbericht(long id)
         {
+            // Zoek weerbericht
             var weerbericht = await _context.Weerberichten.FindAsync(id);
 
+            // Als entry niet bestaat, doe niks
             if (weerbericht == null)
             {
                 return NotFound();
@@ -43,15 +55,22 @@ namespace EducomOpdrachtAPI.Controllers
         }
 
         // GET: api/Weerberichten/bydate/startPeriod/endPeriod
-        // Voor bijgegeven eindperiode
+        /// <summary>
+        /// Haalt alle weerberichten op die binnen de start- en einddatum vallen.
+        /// </summary>
+        /// <param name="startPeriod">Gegeven startdatum, alles voor deze datum wordt weggelaten</param>
+        /// <param name="endPeriod">Gegeven einddatum, alles na deze datum wordt weggelaten</param>
+        /// <returns>Lijst (IEnumerable) met weerberichten</returns>
         [HttpGet("bydate/{startPeriod}/{endPeriod}")]
         public async Task<ActionResult<IEnumerable<Weerbericht>>> GetWeerberichtWithEnd(string startPeriod, string endPeriod)
         {
+            // Converteert datum strings naar daadwerkelijke DateTime
             DateTime startPeriodDate = DateTime.Parse(startPeriod);
             DateTime endPeriodDate = DateTime.Parse(endPeriod);
 
             List<Weerbericht> weerberichten = new List<Weerbericht>();
 
+            // Probeert alle weerberichten op te halen binnen de datumrange
             try
             {
                 weerberichten = await _context.Weerberichten.Where(o => o.Date >= startPeriodDate && o.Date <= endPeriodDate).ToListAsync();
@@ -61,6 +80,7 @@ namespace EducomOpdrachtAPI.Controllers
                 return NotFound();
             }
 
+            // Als er geen zijn gevonden, return een NotFound
             if (weerberichten.Count == 0)
             {
                 return NotFound();
@@ -70,15 +90,23 @@ namespace EducomOpdrachtAPI.Controllers
         }
 
         // GET: api/Weerberichten/bydate/startperiod
-        // Voor standaard +5 dagen eindperiode (zeven dagen is onmogelijk, buienradar heeft maar informatie tot op vijf dagen vooruit)
+        /// <summary>
+        /// Haalt alle weerberichten op die binnen de start- en einddatum vallen, zonder gegeven einddatum.
+        /// </summary>
+        /// <param name="startPeriod">Gegeven startdatum, alles voor deze datum wordt weggelaten</param>
+        /// <returns>Lijst (IEnumerable) met weerberichten</returns>
         [HttpGet("bydate/{startPeriod}")]
         public async Task<ActionResult<IEnumerable<Weerbericht>>> GetWeerberichtWithoutEnd(string startPeriod)
         {
+            // Converteert datum strings naar daadwerkelijke DateTime
             DateTime startPeriodDate = DateTime.Parse(startPeriod);
+
+            // Stel einddatum in als startdatum + 7 dagen
             DateTime endPeriodDate = startPeriodDate.AddDays(7);
 
             List<Weerbericht> weerberichten = new List<Weerbericht>();
 
+            // Probeert alle weerberichten op te halen binnen de datumrange
             try
             {
                 weerberichten = await _context.Weerberichten.Where(o => o.Date >= startPeriodDate && o.Date <= endPeriodDate).ToListAsync();
@@ -88,6 +116,7 @@ namespace EducomOpdrachtAPI.Controllers
                 return NotFound();
             }
 
+            // Als er geen zijn gevonden, return een NotFound
             if (weerberichten.Count == 0)
             {
                 return NotFound();
@@ -97,64 +126,89 @@ namespace EducomOpdrachtAPI.Controllers
         }
 
         // PUT: api/Weerberichten/5
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+        /// <summary>
+        /// Update een weerbericht entry met de gegeven waarden.
+        /// </summary>
+        /// <param name="id">ID van weerbericht</param>
+        /// <param name="weerbericht">Nieuwe waarden voor weerbericht</param>
+        /// <returns>Niks</returns>
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutWeerbericht(long id, Weerbericht weerbericht)
+        public async Task<IActionResult> PutWeerbericht(long id, Weerbericht weerbericht, string requestHeader)
         {
-            if (id != weerbericht.Id)
-            {
-                return BadRequest();
-            }
-
-            var local = _context.Set<Weerbericht>().Local.FirstOrDefault(o => o.Id.Equals((int)id));
-
-            if (local != null)
-            {
-                _context.Entry(local).State = EntityState.Detached;
-            }
-
-            _context.Entry(weerbericht).State = EntityState.Modified;
-
-            try
-            {
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!WeerberichtExists(id))
-                {
-                    return NotFound();
-                }
-                else
-                {
-                    throw;
-                }
-            }
-
-            return NoContent();
-        }
-
-        // POST: api/Weerberichten
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-        [HttpPost]
-        public async Task<ActionResult<Weerbericht>> PostWeerbericht(Weerbericht weerbericht)
-        {
-            // Authenticeer: alleen de console app mag toegang hebben tot POST en PUT
+            // Authenticeer: alleen de console app en website mag toegang hebben tot deze functie
             Authenticator auth = new Authenticator();
             bool authenticated = auth.Authenticate(Request.Headers["Authorization"].ToString());
 
             if (authenticated)
             {
+                // Check of gegeven ID overeen komt met de ID van het weerbericht
+                if (id != weerbericht.Id)
+                {
+                    return BadRequest();
+                }
+
+                // Zoekt in de database of entry bestaat
+                var local = _context.Set<Weerbericht>().Local.FirstOrDefault(o => o.Id.Equals((int)id));
+
+                // Als entry bestaat, Detach de context zodat dit aangepast kan worden
+                if (local != null)
+                {
+                    _context.Entry(local).State = EntityState.Detached;
+                }
+
+                // Pas de entry aan
+                _context.Entry(weerbericht).State = EntityState.Modified;
+
+                try
+                {
+                    // Sla de entry op
+                    await _context.SaveChangesAsync();
+                }
+                catch (DbUpdateConcurrencyException)
+                {
+                    if (!WeerberichtExists(id))
+                    {
+                        return NotFound();
+                    }
+                    else
+                    {
+                        throw;
+                    }
+                }
+            }
+            
+            return NoContent();
+        }
+
+        // POST: api/Weerberichten
+        /// <summary>
+        /// Maak een nieuw weerbericht entry met de gegeven waarden
+        /// </summary>
+        /// <param name="weerbericht">Waarden voor nieuw weerbericht</param>
+        /// <returns>GetWeerbericht van nieuw aangemaakt weerbericht</returns>
+        [HttpPost]
+        public async Task<ActionResult<Weerbericht>> PostWeerbericht(Weerbericht weerbericht)
+        {
+            // Authenticeer: alleen de console app en website mag toegang hebben tot deze functie
+            Authenticator auth = new Authenticator();
+            bool authenticated = auth.Authenticate(Request.Headers["Authorization"].ToString());
+
+            // Als authenticatie succesvol is
+            if (authenticated)
+            {
+                // Check of entry al bestaat
                 if (!_context.Weerberichten.Any(o => o.Date == weerbericht.Date))
                 {
+                    // Nee -> voeg toe (POST)
                     _context.Weerberichten.Add(weerbericht);
                     await _context.SaveChangesAsync();
                 }
                 else
                 {
+                    // Ja -> update (PUT)
                     int id = _context.Weerberichten.First(o => o.Date == weerbericht.Date).Id;
                     weerbericht.Id = id;
-                    await PutWeerbericht((long)id, weerbericht);
+                    await PutWeerbericht((long)id, weerbericht, Request.Headers["Authorization"].ToString());
                 }
 
                 return CreatedAtAction(nameof(GetWeerbericht), new { id = weerbericht.Id }, weerbericht);
@@ -164,31 +218,39 @@ namespace EducomOpdrachtAPI.Controllers
         }
 
         // POST: api/Weerberichten/list
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+        /// <summary>
+        /// Haalt data op uit de Buienradar feed en verwerkt maakt aan (POST) of update (PUT) deze in de database.
+        /// </summary>
+        /// <returns>Niks</returns>
         [HttpGet("list")]
         public async Task<ActionResult<List<Weerbericht>>> GetAndPostWeerberichtList()
         {
-            // Authenticeer: alleen de console app mag toegang hebben tot POST en PUT
+            // Authenticeer: alleen de console app en website mag toegang hebben tot deze functie
             Authenticator auth = new Authenticator();
             bool authenticated = auth.Authenticate(Request.Headers["Authorization"].ToString());
 
+            // Als authenticatie succesvol is
             if (authenticated)
             {
                 Models.Database database = new Models.Database();
                 database.GetWeerberichtenFromFeed();
 
+                // Loop door lijst verkregen van Buienradar
                 foreach (Weerbericht weerbericht in database.weerberichten)
                 {
+                    // Check of entry al bestaat
                     if (!_context.Weerberichten.Any(o => o.Date == weerbericht.Date))
                     {
+                        // Nee: voeg toe (POST)
                         _context.Weerberichten.Add(weerbericht);
                         await _context.SaveChangesAsync();
                     }
                     else
                     {
+                        // Ja -> update (PUT)
                         int id = _context.Weerberichten.First(o => o.Date == weerbericht.Date).Id;
                         weerbericht.Id = id;
-                        await PutWeerbericht((long)id, weerbericht);
+                        await PutWeerbericht((long)id, weerbericht, Request.Headers["Authorization"].ToString());
                     }
                 }
             }
@@ -197,39 +259,69 @@ namespace EducomOpdrachtAPI.Controllers
         }
 
         // DELETE: api/Weerberichten/5
+        /// <summary>
+        /// Delete een specifiek weerbericht
+        /// </summary>
+        /// <param name="id">ID van weerbericht</param>
+        /// <returns>Niks</returns>
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteWeerbericht(long id)
         {
-            var weerbericht = await _context.Weerberichten.FindAsync(id);
-            if (weerbericht == null)
-            {
-                return NotFound();
-            }
+            // Authenticeer: alleen de console app en website mag toegang hebben tot deze functie
+            Authenticator auth = new Authenticator();
+            bool authenticated = auth.Authenticate(Request.Headers["Authorization"].ToString());
 
-            _context.Weerberichten.Remove(weerbericht);
-            await _context.SaveChangesAsync();
+            // Als authenticatie succesvol is
+            if (authenticated)
+            {
+                // Zoek weerbericht entry
+                var weerbericht = await _context.Weerberichten.FindAsync(id);
+
+                // Als entry niet bestaat, doe niks
+                if (weerbericht == null)
+                {
+                    return NotFound();
+                }
+
+                // Als entry bestaat, verwijder en sla op
+                _context.Weerberichten.Remove(weerbericht);
+                await _context.SaveChangesAsync();
+            }
 
             return NoContent();
         }
 
         // DELETE: api/Weerstations
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+        /// <summary>
+        /// Delete alle weerberichten
+        /// </summary>
+        /// <returns>Niks</returns>
         [HttpDelete]
         public async Task<ActionResult<Weerbericht>> DeleteAllWeerbericht()
         {
-            // Authenticeer: alleen de console app mag toegang hebben tot POST en PUT
+            // Authenticeer: alleen de console app en website mag toegang hebben tot deze functie
             Authenticator auth = new Authenticator();
             bool authenticated = auth.Authenticate(Request.Headers["Authorization"].ToString());
 
+            // Als authenticatie succesvol is
             if (authenticated)
             {
+                // Zoek weerbericht entries
                 var weerberichtList = await _context.Weerberichten.ToListAsync();
 
+                // Als entries niet bestaan, doe niks
+                if (weerberichtList.Count == 0)
+                {
+                    return NotFound();
+                }
+
+                // Als entries bestaan, verwijder
                 foreach (Weerbericht weerbericht in weerberichtList)
                 {
                     _context.Weerberichten.Remove(weerbericht);
                 }
 
+                // Sla wijzigingen op
                 await _context.SaveChangesAsync();
             }
