+++ EducomOpdrachtAPI/Controllers/WeerstationsController.cs
@@ -13,6 +13,7 @@ namespace EducomOpdrachtAPI.Controllers
     [Route("api/[controller]")]
     public class WeerstationsController : ControllerBase
     {
+        // Stelt de database context in (Entity Framework)
         private readonly WeerstationContext _context;
 
         public WeerstationsController(WeerstationContext context)
@@ -21,6 +22,10 @@ namespace EducomOpdrachtAPI.Controllers
         }
 
         // GET: api/Weerstations
+        /// <summary>
+        /// Haalt alle weerstations op uit de database.
+        /// </summary>
+        /// <returns>Lijst met weerstations</returns>
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Weerstation>>> GetWeerstations()
         {
@@ -28,11 +33,18 @@ namespace EducomOpdrachtAPI.Controllers
         }
 
         // GET: api/Weerstations/5
+        /// <summary>
+        /// Haalt een specifiek weerstation op.
+        /// </summary>
+        /// <param name="id">ID van weerstation</param>
+        /// <returns>Weerstation</returns>
         [HttpGet("{id}")]
         public async Task<ActionResult<Weerstation>> GetWeerstation(long id)
         {
+            // Zoek weerstation
             var weerstation = await _context.Weerstations.FindAsync(id);
 
+            // Als entry niet bestaat, doe niks
             if (weerstation == null)
             {
                 return NotFound();
@@ -42,15 +54,22 @@ namespace EducomOpdrachtAPI.Controllers
         }
 
         // GET: api/Weerstations/bydate/startPeriod/endPeriod
-        // Voor bijgegeven eindperiode
+        /// <summary>
+        /// Haalt alle weerstation entries op die binnen de start- en einddatum vallen.
+        /// </summary>
+        /// <param name="startPeriod">Gegeven startdatum, alles voor deze datum wordt weggelaten</param>
+        /// <param name="endPeriod">Gegeven einddatum, alles na deze datum wordt weggelaten</param>
+        /// <returns>Lijst (IEnumerable) met weerstations</returns>
         [HttpGet("bydate/{stationId}/{startPeriod}/{endPeriod}")]
         public async Task<ActionResult<IEnumerable<Weerstation>>> GetWeerstationWithEnd(long stationId, string startPeriod, string endPeriod)
         {
+            // Converteert datum strings naar daadwerkelijke DateTime
             DateTime startPeriodDate = DateTime.Parse(startPeriod);
             DateTime endPeriodDate = DateTime.Parse(endPeriod);
 
             List<Weerstation> weerstations = new List<Weerstation>();
 
+            // Probeert alle weerstations op te halen binnen de datumrange
             try
             {
                 weerstations = await _context.Weerstations.Where(o => o.StationId == stationId && o.Date >= startPeriodDate && o.Date <= endPeriodDate).ToListAsync();
@@ -60,6 +79,7 @@ namespace EducomOpdrachtAPI.Controllers
                 return NotFound();
             }
 
+            // Als er geen zijn gevonden, return een NotFound
             if (weerstations.Count == 0)
             {
                 return NotFound();
@@ -69,15 +89,23 @@ namespace EducomOpdrachtAPI.Controllers
         }
 
         // GET: api/Weerstations/bydate/startperiod
-        // Voor standaard +5 dagen eindperiode (zeven dagen is onmogelijk, buienradar heeft maar informatie tot op vijf dagen vooruit)
+        /// <summary>
+        /// Haalt alle weerstation entries op die binnen de start- en einddatum vallen, zonder gegeven einddatum.
+        /// </summary>
+        /// <param name="startPeriod">Gegeven startdatum, alles voor deze datum wordt weggelaten</param>
+        /// <returns>Lijst (IEnumerable) met weerstations</returns> [HttpGet("bydate/{stationId}/{startPeriod}")]
         [HttpGet("bydate/{stationId}/{startPeriod}")]
         public async Task<ActionResult<IEnumerable<Weerstation>>> GetWeerstationWithoutEnd(long stationId, string startPeriod)
         {
+            // Converteert datum strings naar daadwerkelijke DateTime
             DateTime startPeriodDate = DateTime.Parse(startPeriod);
+
+            // Stel einddatum in als startdatum + 7 dagen
             DateTime endPeriodDate = startPeriodDate.AddDays(7);
 
             List<Weerstation> weerstations = new List<Weerstation>();
 
+            // Probeert alle weerstations op te halen binnen de datumrange
             try
             {
                 weerstations = await _context.Weerstations.Where(o => o.StationId == stationId && o.Date >= startPeriodDate && o.Date <= endPeriodDate).ToListAsync();
@@ -87,6 +115,7 @@ namespace EducomOpdrachtAPI.Controllers
                 return NotFound();
             }
 
+            // Als er geen zijn gevonden, return een NotFound
             if (weerstations.Count == 0)
             {
                 return NotFound();
@@ -96,37 +125,55 @@ namespace EducomOpdrachtAPI.Controllers
         }
 
         // PUT: api/Weerstations/5
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+        /// <summary>
+        /// Update een weerstation entry met de gegeven waarden.
+        /// </summary>
+        /// <param name="id">ID van weerstation</param>
+        /// <param name="weerstation">Nieuwe waarden voor weerstation</param>
+        /// <returns>Niks</returns>
         [HttpPut("{id}")]
         public async Task<IActionResult> PutWeerstation(long id, Weerstation weerstation)
         {
-            if (id != weerstation.Id)
-            {
-                return BadRequest();
-            }
+            // Authenticeer: alleen de console app en website mag toegang hebben tot deze functie
+            Authenticator auth = new Authenticator();
+            bool authenticated = auth.Authenticate(Request.Headers["Authorization"].ToString());
 
-            var local = _context.Set<Weerbericht>().Local.FirstOrDefault(o => o.Id.Equals((int)id));
-
-            if (local != null)
+            // Als authenticatie succesvol is
+            if (authenticated)
             {
-                _context.Entry(local).State = EntityState.Detached;
-            }
-
-            _context.Entry(weerstation).State = EntityState.Modified;
-            
-            try
-            {
-                await _context.SaveChangesAsync();
-            }
-            catch (DbUpdateConcurrencyException)
-            {
-                if (!WeerstationExists(id))
+                // Check of gegeven ID overeen komt met de ID van het weerstation
+                if (id != weerstation.Id)
                 {
-                    return NotFound();
+                    return BadRequest();
                 }
-                else
+
+                // Zoekt in de database of entry bestaat
+                var local = _context.Set<Weerbericht>().Local.FirstOrDefault(o => o.Id.Equals((int)id));
+
+                // Als entry bestaat, Detach de context zodat dit aangepast kan worden
+                if (local != null)
                 {
-                    throw;
+                    _context.Entry(local).State = EntityState.Detached;
+                }
+
+                // Pas de entry aan
+                _context.Entry(weerstation).State = EntityState.Modified;
+
+                try
+                {
+                    // Sla de entry op
+                    await _context.SaveChangesAsync();
+                }
+                catch (DbUpdateConcurrencyException)
+                {
+                    if (!WeerstationExists(id))
+                    {
+                        return NotFound();
+                    }
+                    else
+                    {
+                        throw;
+                    }
                 }
             }
 
@@ -134,23 +181,31 @@ namespace EducomOpdrachtAPI.Controllers
         }
 
         // POST: api/Weerstations
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+        /// <summary>
+        /// Maak een nieuw weerstation entry met de gegeven waarden
+        /// </summary>
+        /// <param name="weerstation">Waarden voor nieuw weerstation</param>
+        /// <returns>GetWeerstation van nieuw aangemaakt weerstation</returns>
         [HttpPost]
         public async Task<ActionResult<Weerstation>> PostWeerstation(Weerstation weerstation)
         {
-            // Authenticeer: alleen de console app mag toegang hebben tot POST en PUT
+            // Authenticeer: alleen de console app en website mag toegang hebben tot deze functie
             Authenticator auth = new Authenticator();
             bool authenticated = auth.Authenticate(Request.Headers["Authorization"].ToString());
 
+            // Als authenticatie succesvol is
             if (authenticated)
             {
+                // Check of entry al bestaat
                 if (!_context.Weerstations.Any(o => o.StationId == weerstation.StationId && o.Date == weerstation.Date))
                 {
+                    // Nee -> voeg toe (POST)
                     _context.Weerstations.Add(weerstation);
                     await _context.SaveChangesAsync();
                 }
                 else
                 {
+                    // Ja -> update (PUT)
                     int id = _context.Weerstations.First(o => o.StationId == weerstation.StationId && o.Date == weerstation.Date).Id;
                     weerstation.Id = id;
                     await PutWeerstation((long)id, weerstation);
@@ -163,28 +218,36 @@ namespace EducomOpdrachtAPI.Controllers
         }
 
         // POST: api/Weerstations/list
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+        /// <summary>
+        /// Haalt data op uit de Buienradar feed en verwerkt maakt aan (POST) of update (PUT) deze in de database.
+        /// </summary>
+        /// <returns>Niks</returns>
         [HttpGet("list")]
         public async Task<ActionResult<List<Weerstation>>> GetAndPostWeerstationList()
         {
-            // Authenticeer: alleen de console app mag toegang hebben tot POST en PUT
+            // Authenticeer: alleen de console app en website mag toegang hebben tot deze functie
             Authenticator auth = new Authenticator();
             bool authenticated = auth.Authenticate(Request.Headers["Authorization"].ToString());
 
+            // Als authenticatie succesvol is
             if (authenticated)
             {
                 Models.Database database = new Models.Database();
                 database.GetWeerstationsFromFeed();
 
+                // Loop door lijst verkregen van Buienradar
                 foreach (Weerstation weerstation in database.weerstations)
                 {
+                    // Check of entry al bestaat
                     if (!_context.Weerstations.Any(o => o.StationId == weerstation.StationId && o.Date == weerstation.Date))
                     {
+                        // Nee: voeg toe (POST)
                         _context.Weerstations.Add(weerstation);
                         await _context.SaveChangesAsync();
                     }
                     else
                     {
+                        // Ja -> update (PUT)
                         int id = _context.Weerstations.First(o => o.StationId == weerstation.StationId && o.Date == weerstation.Date).Id;
                         weerstation.Id = id;
                         await PutWeerstation((long)id, weerstation);
@@ -196,39 +259,69 @@ namespace EducomOpdrachtAPI.Controllers
         }
 
         // DELETE: api/Weerstations/5
+        /// <summary>
+        /// Delete een specifiek weerstation
+        /// </summary>
+        /// <param name="id">ID van weerstation</param>
+        /// <returns>Niks</returns>
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteWeerstation(long id)
         {
-            var weerstation = await _context.Weerstations.FindAsync(id);
-            if (weerstation == null)
-            {
-                return NotFound();
-            }
+            // Authenticeer: alleen de console app en website mag toegang hebben tot deze functie
+            Authenticator auth = new Authenticator();
+            bool authenticated = auth.Authenticate(Request.Headers["Authorization"].ToString());
 
-            _context.Weerstations.Remove(weerstation);
-            await _context.SaveChangesAsync();
+            // Als authenticatie succesvol is
+            if (authenticated)
+            {
+                // Zoek weerstation entry
+                var weerstation = await _context.Weerstations.FindAsync(id);
+
+                // Als entry niet bestaat, doe niks
+                if (weerstation == null)
+                {
+                    return NotFound();
+                }
+
+                // Als entry bestaat, verwijder en sla op
+                _context.Weerstations.Remove(weerstation);
+                await _context.SaveChangesAsync();
+            }
 
             return NoContent();
         }
 
         // DELETE: api/Weerstations
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+        /// <summary>
+        /// Delete alle weerstations
+        /// </summary>
+        /// <returns>Niks</returns>
         [HttpDelete]
         public async Task<ActionResult<Weerstation>> DeleteAllWeerstation()
         {
-            // Authenticeer: alleen de console app mag toegang hebben tot POST en PUT
+            // Authenticeer: alleen de console app en website mag toegang hebben tot deze functie
             Authenticator auth = new Authenticator();
             bool authenticated = auth.Authenticate(Request.Headers["Authorization"].ToString());
 
+            // Als authenticatie succesvol is
             if (authenticated)
             {
+                // Zoek weerstation entries
                 var weerstationList = await _context.Weerstations.ToListAsync();
 
+                // Als entries niet bestaan, doe niks
+                if (weerstationList.Count == 0)
+                {
+                    return NotFound();
+                }
+
+                // Als entries bestaan, verwijder
                 foreach (Weerstation weerstation in weerstationList)
                 {
                     _context.Weerstations.Remove(weerstation);
                 }
 
+                // Sla wijzigingen op
                 await _context.SaveChangesAsync();
             }
